# services/core/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact: {}
  description: B 站风格全栈视频平台核心 API（稿件/评论/收藏/用户）
  title: mybilibili Core API
  version: "1.0"
paths:
  /comment/get-up-like:
    get:
      description: 返回指定 UP 主手动点赞过的评论/回复 ID 列表，用于前端展示「UP 主觉得很赞」徽章
      parameters:
      - description: UP 主用户 id
        in: query
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: UP liked comment ids
          schema:
            type: string
      summary: UP 主觉得很赞
      tags:
      - comment
  /comment/list:
    get:
      description: 分页获取稿件评论；支持按 sort=hot|time 切换排序
      parameters:
      - description: 稿件 id
        in: query
        name: manuscriptId
        required: true
        type: integer
      - description: 页码（默认 1）
        in: query
        name: page
        type: integer
      - description: 每页条数（默认 30）
        in: query
        name: page_size
        type: integer
      - description: hot|time（默认 time）
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment list
          schema:
            type: string
        "400":
          description: manuscriptId 缺失
          schema:
            type: string
      summary: 评论列表
      tags:
      - comment
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"